App.jsx'teki autoLogin sağlayan useEffect, loginAction içerisine alınıp export edildi.
